--- a/Procfile.profile
+++ b/Procfile.profile
@@ -1,0 +1,2 @@
+web: bundle exec rails server
+client: cd client && yarn build:profile