--- a/client/.eslintrc.js
+++ b/client/.eslintrc.js
@@ -23,6 +23,9 @@ module.exports = {
         ],
         extensions: ['.js', '.jsx', '.ts', '.tsx'],
       },
+      node: {
+        extensions: ['.js', '.jsx', '.ts', '.tsx'],
+      },
     },
     react: {
       version: 'detect',