--- a/client/app/bundles/course/assessment/question/programming.jsx
+++ b/client/app/bundles/course/assessment/question/programming.jsx
@@ -1,5 +1,5 @@
 import { render } from 'react-dom';
-import Immutable from 'immutable';
+import { fromJS } from 'immutable';
 import ProviderWrapper from 'lib/components/ProviderWrapper';
 import storeCreator from './programming/store';
 import ProgrammingQuestion from './programming/ProgrammingQuestion';
@@ -8,7 +8,7 @@ $(() => {
   const mountNode = document.getElementById('programming-question');
   if (mountNode) {
     const data = mountNode.getAttribute('data');
-    const props = Immutable.fromJS(JSON.parse(data));
+    const props = fromJS(JSON.parse(data));
     const store = storeCreator(props);
 
     const Page = () => (