--- a/client/app/bundles/course/assessment/question/programming/components/OnlineEditor/index.jsx
+++ b/client/app/bundles/course/assessment/question/programming/components/OnlineEditor/index.jsx
@@ -1,4 +1,4 @@
-import Immutable from 'immutable';
+import { Map } from 'immutable';
 
 import PropTypes from 'prop-types';
 import { injectIntl, defineMessages } from 'react-intl';
@@ -22,7 +22,7 @@ const translations = defineMessages({
 });
 
 const propTypes = {
-  data: PropTypes.instanceOf(Immutable.Map).isRequired,
+  data: PropTypes.instanceOf(Map).isRequired,
   actions: PropTypes.object.isRequired,
   isLoading: PropTypes.bool.isRequired,
   autograded: PropTypes.bool.isRequired,