--- a/client/app/bundles/course/assessment/question/programming/reducers/index.jsx
+++ b/client/app/bundles/course/assessment/question/programming/reducers/index.jsx
@@ -1,10 +1,10 @@
-import Immutable from 'immutable';
+import { fromJS } from 'immutable';
 import { combineReducers } from 'redux-immutable';
 import programmingQuestionReducer, {
   initialState as programmingQuestionState,
 } from './programmingQuestionReducer';
 
-export const initialStates = Immutable.fromJS({
+export const initialStates = fromJS({
   programmingQuestion: programmingQuestionState,
 });