--- a/client/app/bundles/course/assessment/skills/pages/SkillsIndex/index.tsx
+++ b/client/app/bundles/course/assessment/skills/pages/SkillsIndex/index.tsx
@@ -143,7 +143,7 @@ const SkillsIndex: FC<Props> = (props) => {
 
   return (
     <>
-      <PageHeader title={intl.formatMessage({ ...translations.skills })} />
+      <PageHeader title={intl.formatMessage(translations.skills)} />
       {isLoading ? (
         <LoadingIndicator />
       ) : (