--- a/client/app/bundles/course/discussion/topics/pages/CommentIndex/index.tsx
+++ b/client/app/bundles/course/discussion/topics/pages/CommentIndex/index.tsx
@@ -197,7 +197,7 @@ const CommentIndex: FC<Props> = (props) => {
         title={
           settings.title
             ? settings.title
-            : intl.formatMessage({ ...translations.comments })
+            : intl.formatMessage(translations.comments)
         }
       />
       {isLoading ? (