--- a/client/app/bundles/course/experience-points/disbursement/pages/DisbursementIndex/index.tsx
+++ b/client/app/bundles/course/experience-points/disbursement/pages/DisbursementIndex/index.tsx
@@ -76,9 +76,7 @@ const DisbursementIndex: FC<Props> = (props) => {
 
   return (
     <>
-      <PageHeader
-        title={intl.formatMessage({ ...translations.disbursements })}
-      />
+      <PageHeader title={intl.formatMessage(translations.disbursements)} />
       {isLoading ? (
         <LoadingIndicator />
       ) : (