--- a/client/app/lib/components/DataTable.jsx
+++ b/client/app/lib/components/DataTable.jsx
@@ -1,4 +1,4 @@
-import { createTheme, ThemeProvider, useTheme } from '@mui/material';
+import { createTheme, ThemeProvider, useTheme } from '@mui/material/styles';
 import MUIDataTable from 'mui-datatables';
 
 const options = {