--- a/client/package.json
+++ b/client/package.json
@@ -162,6 +162,7 @@
     "babel-jest": "^29.0.3",
     "babel-loader": "^8.2.5",
     "babel-plugin-import": "^1.13.5",
+    "babel-plugin-lodash": "^3.3.4",
     "babel-plugin-react-remove-properties": "^0.3.0",
     "css-loader": "^6.7.1",
     "cssnano": "^5.1.13",