--- a/client/package.json
+++ b/client/package.json
@@ -183,7 +183,6 @@
     "eslint-plugin-react-hooks": "^4.6.0",
     "expose-loader": "^4.0.0",
     "fork-ts-checker-webpack-plugin": "^7.2.13",
-    "hard-source-webpack-plugin": "^0.13.1",
     "jest": "^27.5.1",
     "jest-localstorage-mock": "^2.4.22",
     "postcss": "^8.4.16",