--- a/client/webpack.common.js
+++ b/client/webpack.common.js
@@ -1,16 +1,9 @@
-const webpack = require('webpack');
-const path = require('path');
+const { join, resolve } = require('path');
+const { IgnorePlugin, ContextReplacementPlugin } = require('webpack');
 const { WebpackManifestPlugin } = require('webpack-manifest-plugin');
-const HardSourceWebpackPlugin = require('hard-source-webpack-plugin');
 const ForkTsCheckerWebpackPlugin = require('fork-ts-checker-webpack-plugin');
 
-const env = process.env.NODE_ENV || 'development';
-const development = env === 'development';
-const travis = process.env.TRAVIS === 'true';
-
-const config = {
-  mode: env,
-
+module.exports = {
   entry: {
     coursemology: [
       '@babel/polyfill',
@@ -19,26 +12,21 @@ const config = {
       './app/lib/moment-timezone',
     ],
   },
-
   output: {
-    filename: development ? '[name].js' : '[name]-[contenthash].js',
-    path: path.join(__dirname, '..', 'public', 'webpack'),
-    publicPath: '/webpack/',
+    path: join(__dirname, '..', 'public', 'webpack'),
   },
-
   resolve: {
     extensions: ['.js', '.jsx', '.ts', '.tsx'],
     alias: {
-      api: path.resolve('./app/api'),
-      lib: path.resolve('./app/lib'),
-      theme: path.resolve('./app/theme'),
-      types: path.resolve('./app/types'),
-      utilities: path.resolve('./app/utilities'),
-      course: path.resolve('./app/bundles/course'),
-      testUtils: path.resolve('./app/__test__/utils'),
+      api: resolve('./app/api'),
+      lib: resolve('./app/lib'),
+      theme: resolve('./app/theme'),
+      types: resolve('./app/types'),
+      utilities: resolve('./app/utilities'),
+      course: resolve('./app/bundles/course'),
+      testUtils: resolve('./app/__test__/utils'),
     },
   },
-
   optimization: {
     splitChunks: {
       chunks: 'all',
@@ -51,18 +39,14 @@ const config = {
     },
     moduleIds: 'deterministic',
   },
-
   plugins: [
-    new webpack.IgnorePlugin({ resourceRegExp: /__test__/ }),
+    new IgnorePlugin({ resourceRegExp: /__test__/ }),
     new WebpackManifestPlugin({
       publicPath: '/webpack/',
       writeToFileEmit: true,
     }),
     // Do not require all locales in moment
-    new webpack.ContextReplacementPlugin(
-      /moment\/locale$/,
-      /^\.\/(en-.*|zh-.*)$/,
-    ),
+    new ContextReplacementPlugin(/moment\/locale$/, /^\.\/(en-.*|zh-.*)$/),
     new ForkTsCheckerWebpackPlugin({
       typescript: {
         diagnosticOptions: {
@@ -73,7 +57,6 @@ const config = {
       },
     }),
   ],
-
   module: {
     rules: [
       {
@@ -89,12 +72,19 @@ const config = {
       },
       {
         test: /\.css$/,
+        use: ['style-loader', 'css-loader'],
+        include: [resolve(__dirname, 'node_modules/rc-slider/assets')],
+      },
+      {
+        test: /\.css$/,
         use: ['style-loader', 'css-loader', 'postcss-loader'],
+        include: [resolve(__dirname, 'app/theme/index.css')],
       },
       {
         test: /\.scss$/,
         use: ['style-loader', 'css-loader', 'sass-loader'],
-        include: [path.resolve(__dirname, 'app/lib/styles')],
+        include: [resolve(__dirname, 'app/lib/styles')],
+        exclude: /node_modules/,
       },
       {
         test: /\.scss$/,
@@ -111,7 +101,7 @@ const config = {
           },
           'sass-loader',
         ],
-        exclude: [/node_modules/, path.resolve(__dirname, 'app/lib/styles')],
+        exclude: [/node_modules/, resolve(__dirname, 'app/lib/styles')],
       },
       {
         test: require.resolve('jquery'),
@@ -130,26 +120,3 @@ const config = {
     ],
   },
 };
-
-if (development) {
-  const devServerPort = 8080;
-  config.devServer = {
-    port: devServerPort,
-    headers: { 'Access-Control-Allow-Origin': '*' },
-  };
-  config.output.publicPath = `//localhost:${devServerPort}/webpack/`;
-  config.devtool = 'eval-cheap-module-source-map';
-} else {
-  console.log(`\nWebpack ${env} build for Rails...`); // eslint-disable-line no-console
-}
-
-// Only enable HardSourceWebpackPlugin in Travis
-if (travis) {
-  config.plugins.push(
-    new HardSourceWebpackPlugin({
-      cacheDirectory: path.join(__dirname, 'hard-source-cache/[confighash]'),
-    }),
-  );
-}
-
-module.exports = config;