--- a/client/webpack.prod.js
+++ b/client/webpack.prod.js
@@ -1,5 +1,4 @@
 const { merge } = require('webpack-merge');
-const HardSourceWebpackPlugin = require('hard-source-webpack-plugin');
 
 const common = require('./webpack.common');
 
@@ -9,11 +8,6 @@ module.exports = merge(common, {
     filename: '[name]-[contenthash].js',
     publicPath: '/webpack/',
   },
-  plugins: [
-    new HardSourceWebpackPlugin({
-      cacheDirectory: join(__dirname, 'hard-source-cache/[confighash]'),
-    }),
-  ],
   optimization: {
     usedExports: true,
   },